third_party/blink/web_tests/platform/generic/css3/calc/font-size-fractional-expected.txt

Tests that CSS3 calc() can be used with the font property using fractional sizes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById("calc-reduce"), null).lineHeight is "24px"
PASS getComputedStyle(document.getElementById("calc-reduce"), null).fontSize is "9.6px"
PASS successfullyParsed is true

TEST COMPLETE
